Refactor post page details file structure
