+++ components/postPage/details/dimensions.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Row, Col } from "antd";
 
-import { ContainerNoOutline, Title, ItemKey, ItemValue } from "../style";
+import { ContainerNoOutline, Title, ItemKey, ItemValue } from "./style";
 
 const index = ({ post }) => {
   return (
