+++ components/postPage/similarPosts/PostCard.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const PostCard = () => {
+  return (
+    <div>
+      <div>post card</div>
+    </div>
+  );
+};
+
+export default PostCard;
