+++ components/postPage/similarPosts/style.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const Title = styled.div`
+  font-size: 18px;
+  font-weight: bold;
+  color: #00458a;
+`;
+
+export const ContainerOutline = styled.div`
+  border-top: 2px solid #00458a;
+  padding: 5px;
+`;
